♻️ refactor: extract validation and category logic to utils
Move validateFilename() to src/utils/validateFilename.ts and
categoryName()/categoryMimes() (with MIME_CATEGORIES/ALL_OFFICE_MIMES) to
src/utils/fileCategories.ts. Behaviour-preserving — OfficeOverview.vue
calls the same functions, just imported instead of defined inline.

filterFiles (the Mine-filter logic) is extracted separately, with
#46's external-storage exclusion fix applied.
